refactor(xit): separate action from burn
Move generation of XIT JSON into own composable and separate from burn drawer.
